Extract SMB and PostgreSQL optional sessions into their own mixins
