+++ modules/auxiliary/scanner/smb/smb_enumshares.rb
@@ -14,7 +14,7 @@ class MetasploitModule < Msf::Auxiliary
   include Msf::Auxiliary::Report
   include Msf::Auxiliary::Scanner
 
-  include Msf::OptionalSession
+  include Msf::OptionalSession::SMB
 
   def initialize(info = {})
     super(
