+++ modules/exploits/windows/smb/psexec.rb
@@ -21,7 +21,7 @@ class MetasploitModule < Msf::Exploit::Remote
   include Msf::Exploit::EXE
   include Msf::Exploit::WbemExec
   include Msf::Auxiliary::Report
-  include Msf::OptionalSession
+  include Msf::OptionalSession::SMB
 
   def initialize(info = {})
     super(update_info(info,
